Handle nullable LHS in "??".
We can strip off its nullity when determining the result type of the
"??" expression since the LHS will only be used when it is not null.

This makes it nicer to use "??" in places where you want to check for
and handle a nullable type:

String? maybe = ...
String definitely = maybe ?? "default"; // OK now.